rasa/requirements.txt
rasa==3.6.15
rasa-sdk==3.6.2
requests==2.31.0
python-dotenv==1.0.0 